Fix url pathing and redirect user back to MauHienTruong/Index after committing create YeuCauLayMau
