--- Source-Codes/IERSystem/IERSystem/Areas/Administrator/Controllers/RequestController.cs
@@ -1,159 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Net;
-using System.Web;
-using System.Web.Mvc;
-using IERSystem.Areas.Administrator.Models;
-using System.IO;
-using Newtonsoft.Json;
-
-namespace IERSystem.Areas.Administrator.Controllers
-{
-  
-    public class RequestController : Controller
-    {
-        private IERSystemDBContext db = new IERSystemDBContext();
-
-        // GET: /Administrator/Request/
-        public async Task<ActionResult> Index()
-        {
-            return View(await db.Requests.ToListAsync());
-        }
-
-        // GET: /Administrator/Request/Details/5
-        public async Task<ActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Request request = await db.Requests.FindAsync(id);
-            if (request == null)
-            {
-                return HttpNotFound();
-            }
-            return View(request);
-        }
-
-        // GET: /Administrator/Request/Create
-        public ActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: /Administrator/Request/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create(
-            [Bind(Include="MaDon,TenKhachHang,TenDaiDien,DiaChiLayMau,MaSoThue,SoDienThoai,SoFax,NgayTaoHD,NgayDuKienTraMau, MauLayHienTruong")]
-            RequestInputModel inputRequest)
-        {
-            //bool isOk = false;
-            //string errMsg = "";
-            if (ModelState.IsValid)
-            {
-                //    db.Requests.Add(request);
-                //    await db.SaveChangesAsync();
-                //    return RedirectToAction("Index");
-                //return Json(new { isOk, errMsg });
-            }
-
-            return View(inputRequest);
-        }
-
-        // POST: /Administrator/Request/CreateNew
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        public async Task<JsonResult> CreateNew(RequestInputModel inputRequest) 
-        {
-            bool isOk = false;
-            string errMsg = "";
-            if (ModelState.IsValid) {
-                db.Requests.Add(inputRequest.ToModel());
-                try {
-                    await db.SaveChangesAsync();
-                    Console.WriteLine("OK");
-                    isOk = true;
-                } catch (System.Data.Entity.Infrastructure.DbUpdateException e) {
-                    Console.WriteLine(e.Message);
-                    isOk = false;
-                }
-            }
-
-            return Json(new { isOk, errMsg });
-        }
-
-        // GET: /Administrator/Request/Edit/5
-        public async Task<ActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Request request = await db.Requests.FindAsync(id);
-            if (request == null)
-            {
-                return HttpNotFound();
-            }
-            return View(request);
-        }
-
-        // POST: /Administrator/Request/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include="Id,MaDon,TenKhachHang,TenDaiDien,DiaChiLayMau,DiaChiKhachHang,MaSoThue,SoDienThoai,SoFax,NgayTaoHD,NgayDuKienTraMau,PhuongPhapLayMau,TenTieuChuanDoiChieu,PhiThiNghiemTamTinh,TienKhachHangTraTruoc,DaGuiMau")] Request request)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(request).State = EntityState.Modified;
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
-            }
-            return View(request);
-        }
-
-        // GET: /Administrator/Request/Delete/5
-        public async Task<ActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Request request = await db.Requests.FindAsync(id);
-            if (request == null)
-            {
-                return HttpNotFound();
-            }
-            return View(request);
-        }
-
-        // POST: /Administrator/Request/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> DeleteConfirmed(int id)
-        {
-            Request request = await db.Requests.FindAsync(id);
-            db.Requests.Remove(request);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
-        }
-    }
-}
