+++ Source-Codes/IERSystem/IERSystem/Areas/HopDongLayMau/Controllers/APIController.cs
@@ -1,0 +1,51 @@
+﻿using IERSystem.Areas.Administrator.Models;
+using IERSystem.Areas.HopDongLayMau.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Web;
+using System.Web.Mvc;
+
+namespace IERSystem.Areas.HopDongLayMau.Controllers
+{
+    public class APIController : Controller
+    {
+        private IERSystemDBContext db = new IERSystemDBContext();
+
+        // GET: HopDongLayMau/API
+        public ActionResult Index()
+        {
+            return View();
+        }
+
+        // POST: /HopDongLayMau/API/Create
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        public async Task<JsonResult> Create(YeuCauLayMauInputModel inputRequest) {
+            bool isOk = false;
+            string errMsg = "";
+            if (ModelState.IsValid) {
+                db.Requests.Add(inputRequest.ToModel());
+                try {
+                    await db.SaveChangesAsync();
+                    Console.WriteLine("OK");
+                    isOk = true;
+                } catch (System.Data.Entity.Infrastructure.DbUpdateException e) {
+                    Console.WriteLine(e.Message);
+                    isOk = false;
+                }
+            }
+
+            return Json(new { isOk, errMsg });
+        }
+
+        protected override void Dispose(bool disposing) {
+            if (disposing) {
+                db.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+    }
+}
