+++ Source-Codes/IERSystem/IERSystem/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="Antlr" version="3.4.1.9004" targetFramework="net45" />
   <package id="bootstrap" version="3.0.0" targetFramework="net45" />
-  <package id="EntityFramework" version="6.0.0" targetFramework="net45" />
+  <package id="EntityFramework" version="6.1.3" targetFramework="net451" />
   <package id="jQuery" version="1.10.2" targetFramework="net45" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net45" />
   <package id="Microsoft.AspNet.Identity.Core" version="1.0.0" targetFramework="net45" />
