+++ Source-Codes/IERSystem/packages/EntityFramework.6.1.3/content/Web.config.transform
@@ -1,0 +1,5 @@
+﻿<configuration>
+    <configSections>
+        <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 -->
+    </configSections>
+</configuration>
